# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: BookMyShoot API
  description: Marketplace API for booking photographers, videographers and editors in Sri Lanka
  version: 1.0.0
servers:
  - url: https://api.bookmyshoot.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /services:
    get:
      summary: Search services
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: rating
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
...
  /bookings:
    post:
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_id:
                  type: integer
                package_id:
                  type: integer
                date:
                  type: string
                  format: date
                start_time:
                  type: string
                end_time:
                  type: string
              required:
                - service_id
                - date
                - start_time
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Bad request
        '409':
          description: Time slot not available
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        pro_id:
          type: integer
        title:
          type: string
        category:
          type: string
        base_price:
          type: number
        duration_minutes:
          type: integer
        is_active:
          type: boolean
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        pro_id:
          type: integer
        service_id:
          type: integer
        package_id:
          type: integer
        date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, rejected]
        total_price:
          type: number
        deposit_paid:
          type: number
        created_at:
          type: string
          format: date-time